add TYPE_CHECKING for pandas import
